Trim default-valued build settings after Starlark transition
By convention, BuildOptions should not contain Starlark build settings
that are set to their default values. However, if a Starlark transition
has a build setting as an input, the value of that setting has to be
added to the input BuildOptions temporarily so that it can be read by
the transition.

Before this commit, these build settings remained in the transition and
could cause action conflicts. With this commit, all build settings that
were added with their default values are iterated through and removed if
they are still set to their defaults post-transition.

Fixes #12888.